requirements.txt
boto3>=1.26.0
slack-sdk>=3.21.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
